requirement.txt
mysql-connector-python
numpy
pillow
PyAutoGUI
opencv-python
threading2
deepface